Added support to track type of tasks such as todo, deadline and event.
created 3 more classes for todo, deadline and event. each were inherited from task class and each overrides to string accordingly. for deadline and event, split function was used to split the sentence into arrays of description of tasks and the time information. during "list", will list out the task type, if its marked/unmarked, description and the timing if applicable. Also changed the if-else structure of runChat slightly.